Fix a bug of wrong line number
This patch fixed the wrong positions of non-terminals due to block
comments. The escape \n automatically updates yylineno if its rule
is defined explicitly (without action code) in *.l files. As error
recovery is left for future work, not all syntax errors are reported,
the parser exits when the first error occurs.